transition to new classes merge 1#
