+++ udsp/client/configCryptography.js
@@ -1,0 +1,72 @@
+import { hasValue, has, intersection } from '@universalweb/acid';
+import { success } from '#logs';
+import { getAlgorithm, processPublicKey } from '../cryptoMiddleware/index.js';
+export async function configCryptography() {
+	// console.log(this.cryptography);
+	const { destination, } = this;
+	const {
+		encryptConnectionId,
+		publicKeyAlgorithm,
+	} = destination;
+	if (!destination.cipherSuites) {
+		destination.cipherSuites = this.cipherSuites;
+	}
+	if (destination.cipherSuites) {
+		if (!has(destination.cipherSuites, this.cipherSuiteName)) {
+			console.log('Default ciphersuite not available');
+			this.cipherSuiteName = intersection(this.cipherSuites, destination.cipherSuites)[0];
+			if (!this.cipherSuiteName) {
+				console.log('No matching cipher suite found.');
+				return false;
+			}
+		}
+	}
+	this.publicKeyCryptography = getAlgorithm(publicKeyAlgorithm, this.version);
+	this.cipherSuite = getAlgorithm(this.cipherSuiteName, this.version);
+	console.log(this.cipherSuiteName);
+	if (destination.boxCryptography) {
+		this.boxCryptography = getAlgorithm(destination.boxCryptography, this.version);
+	}
+	this.compression = destination.compression;
+	this.headerCompression = destination.headerCompression;
+	if (destination.autoLogin && this.autoLogin) {
+		this.autoLogin = true;
+	}
+	if (!this.keypair) {
+		this.keypair = this.cipherSuite.keypair();
+		success(`Created Connection Keypair`);
+	}
+	if (!this.encryptionKeypair) {
+		this.encryptionKeypair = this.keypair;
+	}
+	const convertSignKeypairToEncryptionKeypair = processPublicKey(this.destination);
+	if (convertSignKeypairToEncryptionKeypair) {
+		this.destination.encryptionKeypair = convertSignKeypairToEncryptionKeypair;
+	}
+	await this.setSessionKeys();
+	if (encryptConnectionId) {
+		const {
+			server: encryptServerCid,
+			client: encryptClientCid,
+			keypair: connectionIdKeypair
+		} = encryptConnectionId;
+		let encryptServer = hasValue(encryptServerCid);
+		let encryptClient = hasValue(encryptClientCid);
+		if (!encryptServer && !encryptClient) {
+			encryptServer = true;
+			encryptClient = true;
+		}
+		if (encryptServer) {
+			this.encryptServerConnectionId = true;
+			if (connectionIdKeypair) {
+				this.destination.connectionIdKeypair = connectionIdKeypair;
+			} else {
+				this.destination.connectionIdKeypair = this.destination.encryptionKeypair;
+			}
+		}
+		if (encryptClient) {
+			this.encryptClientConnectionId = true;
+		}
+		console.log(`Encrypt Connection ID Server ${encryptServer} Client ${encryptClient}`);
+	}
+}
